Toggle signed-in state in the top nav
