--- themes/default/assets/js/marketing.js
@@ -1,1 +1,0 @@
-(()=>{"use strict";var __webpack_modules__={"./src/scss/_marketing.scss":(__unused_webpack_module,__webpack_exports__,__webpack_require__)=>{eval("__webpack_require__.r(__webpack_exports__);\n// extracted by mini-css-extract-plugin\n\n\n//# sourceURL=webpack://theme/./src/scss/_marketing.scss?")},"./src/ts/marketing.ts":(__unused_webpack_module,__webpack_exports__,__webpack_require__)=>{eval('__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _scss_marketing_scss__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../scss/_marketing.scss */ "./src/scss/_marketing.scss");\n\n\n\n//# sourceURL=webpack://theme/./src/ts/marketing.ts?')}},__webpack_module_cache__={};function __webpack_require__(_){var e=__webpack_module_cache__[_];if(void 0!==e)return e.exports;var r=__webpack_module_cache__[_]={exports:{}};return __webpack_modules__[_](r,r.exports,__webpack_require__),r.exports}__webpack_require__.r=_=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(_,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(_,"__esModule",{value:!0})};var __webpack_exports__=__webpack_require__("./src/ts/marketing.ts")})();
