+++ themes/default/layouts/partials/top-nav-user-toggle.html
@@ -1,0 +1,4 @@
+<pulumi-user-toggle>
+    <span slot="signed-in">Pulumi Cloud</span>
+    <span slot="signed-out">Sign In</span>
+</pulumi-user-toggle>
