+++ themes/default/theme/src/scss/components/_user-toggle.scss
@@ -1,0 +1,5 @@
+pulumi-user-toggle {
+    > span[slot] {
+        @apply invisible;
+    }
+}
