+++ themes/default/theme/src/scss/main.scss
@@ -67,7 +67,8 @@
 @import "components/swiper";
 @import "components/slot-machine";
 @import "components/pulumi-tertiary-nav";
-@import "components/pulumi-ai";
+@import "components/ai";
+@import "components/user-toggle";
 @import "components/tabs";
 
 h1,
