+++ themes/default/theme/stencil/src/components/ai/ai.scss
@@ -1,0 +1,3 @@
+pulumi-ai {
+    display: block;
+}
