--- themes/default/theme/stencil/src/components/ai/pulumi-ai.scss
@@ -1,3 +1,0 @@
-pulumi-ai {
-    display: block;
-}
